+++ CHANGELOG.md
@@ -7,14 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.4] - 2026-02-10
 
+### Added
+
+- Added `enableStrictSecretScanning` setting (and corresponding `toggleStrictSecretScanning` command) that, when enabled, makes secret scanning skip files that don't have a common secret indicator (like "API_SECRET").
+
 ### Changed
 
 - GitGerbil will now wait for up to 5 seconds to detect a git repo when activating instead of immediately failing.
 - File extensions like `.test.ts` will be correctly detected as `.ts` files now and scanned if the base extension is in the list of scanned file types.
-- Replaced `enable` and `disable` commands (i.e. `gitgerbil.enableSecretScanning`) with `toggle` commands.
 - Submitting an empty field when running `gitgerbil.setScannedFileTypes` will now reset to the default list of scanned file types instead of an empty list.
 - Updated the README to mention `gitgerbil-ignore-file`.
 - `.env.example` files will no longer be flagged by file path scanning.
+
+### Removed
+
+- Replaced `enable` and `disable` commands (i.e. `gitgerbil.enableSecretScanning`) with `toggle` commands.
+  - `gitgerbil.toggleFilePathScanning`
+  - `gitgerbil.toggleSecretScanning`
+  - `gitgerbil.toggleStrictSecretScanning`
+  - `gitgerbil.toggleCommentScanning`
 
 ## [0.1.3] - 2026-02-08